+++ zemfrog/extensions/sqlalchemy.py
@@ -1,0 +1,7 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+
+
+def init_app(app: Flask):
+    db = SQLAlchemy()
+    db.init_app(app)