+++ zemfrog/generator.py
@@ -1,4 +1,5 @@
 import os
+
 from flask.globals import current_app
 from jinja2 import Template
 
@@ -137,7 +138,7 @@ def g_api(name: str):
     """
 
     validate_module_name(name)
-    api_dir = os.path.join(current_app.root_path, "api")
+    api_dir = os.path.join(current_app.root_path, "apis")
     old_filename = get_template("api", "name.py")
     with open(old_filename) as fp:
         old_data = fp.read()
@@ -173,16 +174,13 @@ def g_api_crud(name: str):
         idx = len(import_name)
         src_model = src_model[idx:]
 
-    api_dir = os.path.join(current_app.root_path, "api")
+    api_dir = os.path.join(current_app.root_path, "apis")
     old_filename = get_template("crud", "name.py")
     with open(old_filename) as fp:
         old_data = fp.read()
         py_t = Template(old_data)
         new_data = py_t.render(
-            name=name,
-            url_prefix=name,
-            src_model=src_model,
-            main_app=main_app,
+            name=name, url_prefix=name, src_model=src_model, main_app=main_app
         )
 
     new_filename = os.path.join(api_dir, name + ".py")