+++ zemfrog/loaders/blueprint.py
@@ -1,4 +1,4 @@
-from flask import Flask, Blueprint
+from flask import Blueprint, Flask
 
 from ..helper import get_import_name, import_attr