+++ zemfrog/loaders/openapi.py
@@ -1,9 +1,11 @@
-from flask import Flask, Blueprint
-from flask_apispec import FlaskApiSpec
 from importlib import import_module
 
+from flask import Flask
+
+from zemfrog.globals import apispec as docs
+
 from ..decorators import api_doc
-from ..helper import get_import_name, import_attr
+from ..helper import get_import_name
 
 
 def loader(app: Flask):
@@ -15,7 +17,6 @@ def loader(app: Flask):
         return
 
     import_name = get_import_name(app)
-    docs: FlaskApiSpec = import_attr(import_name + "extensions.apispec.docs")
     urls = import_module(import_name + "urls")
     docs_params = urls.docs
     routes = urls.routes
@@ -25,7 +26,7 @@ def loader(app: Flask):
         docs.register(view)
 
     apis = app.config.get("APIS", [])
-    api_prefix = "api."
+    api_prefix = "apis."
     for name in apis:
         if not name.startswith(api_prefix):
             name = api_prefix + name