+++ zemfrog/templates/project/auth/views.py
@@ -4,7 +4,7 @@ from flask_apispec import use_kwargs, marshal_with
 from jwt import DecodeError, ExpiredSignatureError
 from marshmallow import fields
 from werkzeug.security import generate_password_hash, check_password_hash
-from zemfrog.decorators import auto_status_code, authenticate
+from zemfrog.decorators import http_code, authenticate
 from zemfrog.helper import db_add, db_update, db_commit, get_mail_template, get_user_roles
 from zemfrog.models import (
     DefaultResponseSchema,
@@ -15,9 +15,9 @@ from zemfrog.models import (
     RequestPasswordResetSchema,
 )
 
-from {{ "" if main_app else ".." }}extensions.marshmallow import ma
+from zemfrog.globals import ma
 from {{ "" if main_app else ".." }}models.user import User, Log, Role, Permission
-from {{ "" if main_app else ".." }}tasks.email import send_email
+from {{ "" if main_app else ".." }}tasks.mail import send_email
 
 
 class PermissionSchema(ma.SQLAlchemyAutoSchema):
@@ -54,7 +54,7 @@ def user_detail():
 @use_kwargs(LoginSchema(), location="form")
 @marshal_with(DefaultResponseSchema, 404)
 @marshal_with(LoginSuccessSchema, 200)
-@auto_status_code
+@http_code
 def login(**kwds):
     """
     Login and get access token.
@@ -74,13 +74,13 @@ def login(**kwds):
         access_token = create_access_token(email, user_claims=claims)
         return {"access_token": access_token}
 
-    return {"reason": "Incorrect email or password.", "status_code": 404}
+    return {"message": "Incorrect email or password.", "code": 404}
 
 
 @use_kwargs(RegisterSchema(), location="form")
 @marshal_with(DefaultResponseSchema, 200)
 @marshal_with(DefaultResponseSchema, 403)
-@auto_status_code
+@http_code
 def register(**kwds):
     """
     Register an account.
@@ -112,24 +112,24 @@ def register(**kwds):
                 )
                 msg = get_mail_template("register.html", token=token)
                 send_email.delay("Registration", html=msg, recipients=[email])
-                reason = "Successful registration."
+                message = "Successful registration."
                 status_code = 200
             else:
-                reason = "Username and password are required."
+                message = "Username and password are required."
                 status_code = 403
         else:
-            reason = "Email already exists."
+            message = "Email already exists."
             status_code = 403
     else:
-        reason = "Email required."
+        message = "Email required."
         status_code = 403
 
-    return {"reason": reason, "status_code": status_code}
+    return {"message": message, "code": status_code}
 
 
 @marshal_with(DefaultResponseSchema, 200)
 @marshal_with(DefaultResponseSchema, 403)
-@auto_status_code
+@http_code
 def confirm_account(token):
     """
     Confirm account.
@@ -143,7 +143,7 @@ def confirm_account(token):
 
         user = User.query.filter_by(email=email).first()
         if user and not user.confirmed:
-            reason = "Confirmed."
+            message = "Confirmed."
             status_code = 200
             db_update(user, confirmed=True, confirmed_at=datetime.utcnow())
 
@@ -151,17 +151,17 @@ def confirm_account(token):
             raise DecodeError
 
     except DecodeError:
-        reason = "Invalid token."
+        message = "Invalid token."
         status_code = 403
 
-    return {"reason": reason, "status_code": status_code}
+    return {"message": message, "code": status_code}
 
 
 @use_kwargs(RequestPasswordResetSchema(), location="form")
 @marshal_with(DefaultResponseSchema, 200)
 @marshal_with(DefaultResponseSchema, 404)
 @marshal_with(DefaultResponseSchema, 403)
-@auto_status_code
+@http_code
 def request_password_reset(**kwds):
     """
     Request a password reset.
@@ -171,10 +171,10 @@ def request_password_reset(**kwds):
     if email:
         user = User.query.filter_by(email=email).first()
         if not user:
-            reason = "User not found."
+            message = "User not found."
             status_code = 404
         else:
-            reason = "A password reset request has been sent."
+            message = "A password reset request has been sent."
             status_code = 200
             token = create_access_token(
                 email,
@@ -189,16 +189,16 @@ def request_password_reset(**kwds):
             user.logs.append(log)
             db_commit()
     else:
-        reason = "Email required."
+        message = "Email required."
         status_code = 403
 
-    return {"reason": reason, "status_code": status_code}
+    return {"message": message, "code": status_code}
 
 
 @marshal_with(DefaultResponseSchema, 200)
 @marshal_with(DefaultResponseSchema, 401)
 @marshal_with(DefaultResponseSchema, 403)
-@auto_status_code
+@http_code
 def confirm_password_reset_token(token):
     """
     Validate password reset token.
@@ -212,20 +212,20 @@ def confirm_password_reset_token(token):
 
         user = User.query.filter_by(email=email).first()
         if user:
-            reason = "Valid token."
+            message = "Valid token."
             status_code = 200
         else:
             raise DecodeError
 
     except DecodeError:
-        reason = "Invalid token."
+        message = "Invalid token."
         status_code = 401
 
     except ExpiredSignatureError:
-        reason = "Token expired."
+        message = "Token expired."
         status_code = 403
 
-    return {"reason": reason, "status_code": status_code}
+    return {"message": message, "code": status_code}
 
 
 @use_kwargs(PasswordResetSchema(), location="form")
@@ -233,7 +233,7 @@ def confirm_password_reset_token(token):
 @marshal_with(DefaultResponseSchema, 403)
 @marshal_with(DefaultResponseSchema, 401)
 @marshal_with(DefaultResponseSchema, 404)
-@auto_status_code
+@http_code
 def password_reset(token, **kwds):
     """
     Reset user password.
@@ -252,18 +252,18 @@ def password_reset(token, **kwds):
             log = Log(updated_at=datetime.utcnow())
             user.logs.append(log)
             db_update(user, password=passw)
-            reason = "Successfully change password."
+            message = "Successfully change password."
             status_code = 200
         else:
-            reason = "User not found."
+            message = "User not found."
             status_code = 404
 
     except DecodeError:
-        reason = "Invalid token."
+        message = "Invalid token."
         status_code = 401
 
     except ExpiredSignatureError:
-        reason = "Token expired."
+        message = "Token expired."
         status_code = 403
 
-    return {"reason": reason, "status_code": status_code}
+    return {"message": message, "code": status_code}