+++ zemfrog/templates/project/models/user.py
@@ -1,5 +1,6 @@
-from {{ "" if main_app else ".." }}extensions.sqlalchemy import db
+from zemfrog.globals import db
 from zemfrog.mixins import UserMixin, RoleMixin, PermissionMixin, LogMixin
+
 
 class User(UserMixin, db.Model):
     pass